added tests for id algorithm
